+++ src/gui/panes/mod.rs
@@ -1,0 +1,5 @@
+use crate::gui::packet_table::PacketTable;
+
+pub mod drill_down;
+pub mod graph;
+pub mod payload;
